feat: using delta to work with double in the calculator

+++ src/main/java/level3/Calculator.java
@@ -5,19 +5,19 @@ public class Calculator {
     public Calculator() {
     }
 
-    public static float sum(float a, float b) {
+    public static double sum(double a, double b) {
         return a + b;
     }
 
-    public static float substraction(float a, float b) {
+    public static double substraction(double a, double b) {
         return a - b;
     }
 
-    public static float multiplication(float a, float b) {
+    public static double multiplication(double a, double b) {
         return a * b;
     }
 
-    public static float division(float a, float b) {
+    public static double division(double a, double b) {
         return a / b;
     }
 }
